Fix cu_seqlens_k dtype: cumsum upcasts int32 to int64, cast back
torch.cumsum() on int32 tensors produces int64 output, but
flash_attn_varlen_func requires int32 for cu_seqlens_k.
Added .to(torch.int32) after cumsum to fix the dtype mismatch.

+++ prefix-sharing/prefix_sharing/integrations/verl_attention.py
@@ -246,7 +246,7 @@ def _make_verl_attention_patch(original_forward: Any) -> Any:
             cu_seqlens_k = torch.tensor(
                 [0] + [prefix_len + suffix_len] * N,
                 device=hidden_states.device, dtype=torch.int32,
-            ).cumsum(0)
+            ).cumsum(0).to(torch.int32)
 
             input_dtype = query_states.dtype
             if input_dtype == torch.float32:
